separate callModal from dag.js
